feat(buttons): Index Page buttons
Adding buttons from a resource in the index page
