+++ src/Buttons/IndexPage/ImportButton.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace MoonShine\Buttons\IndexPage;
+
+use MoonShine\ActionButtons\ActionButton;
+use MoonShine\Components\FormBuilder;
+use MoonShine\Fields\File;
+use MoonShine\Handlers\ImportHandler;
+use MoonShine\Resources\ModelResource;
+
+final class ImportButton
+{
+    public static function for(ModelResource $resource, ImportHandler $import): ActionButton
+    {
+        return ActionButton::make(
+            $import->label(),
+            '#'
+        )
+            ->customAttributes(['class' => 'btn btn-primary'])
+            ->icon($import->iconValue())
+            ->inOffCanvas(
+                fn () => $import->label(),
+                fn (): FormBuilder => FormBuilder::make(
+                    $resource->route('handler', query: ['handlerUri' => $import->uriKey()])
+                )
+                    ->fields([
+                        File::make(column: $import->getInputName())->required(),
+                    ])
+                    ->submit(__('moonshine::ui.confirm'))
+            );
+    }
+}
